fixed step 3 and moved to float count
git-tfs-id: [https://nos4a2.visualstudio.com]$/Private/BarcodePostprocessing;C277

+++ BarcodePostprocessingWPF/Model/InventoryItem.cs
@@ -6,7 +6,7 @@
     {
         public string Barcode { get; set; }
 
-        public int Count { get; set; }
+        public double Count { get; set; }
 
         public string InternalCode { get; set; }
     }
